Move public headers into directory structure mimicking installed structure
This facilitates building applications linking to srt without installing
the libraries (i.e.: projects that build libsrt as a subproject).
