+++ apps/socketoptions.hpp
@@ -15,7 +15,7 @@
 #include <map>
 #include <set>
 #include <vector>
-#include "../srtcore/srt.h" // Devel path
+#include "srt.h" // Devel path
 
 #ifdef _WIN32
 #include "winsock2.h"
