+++ include/srt/filelist.maf
@@ -1,0 +1,8 @@
+PUBLIC HEADERS
+srt.h
+logging_api.h
+access_control.h
+
+PROTECTED HEADERS
+platform_sys.h
+udt.h
